[maven-release-plugin]  copy for tag syncope-0.7.1

+++ console/src/main/java/org/syncope/console/pages/Resources.java
@@ -384,9 +384,16 @@ public class Resources extends BasePage {
 
                     @Override
                     public void onClick(final AjaxRequestTarget target) {
+
                         try {
-                            connectorRestClient.delete(connectorTO.getId());
-                            info(getString("operation_succeded"));
+
+                            if (!checkDeleteIsForbidden(connectorTO)) {
+                                connectorRestClient.delete(connectorTO.getId());
+                                info(getString("operation_succeded"));
+                            } else {
+                                error(getString("delete_error"));
+                            }
+
                         } catch (SyncopeClientCompositeErrorException e) {
                             error(getString("operation_error"));
 
@@ -490,6 +497,24 @@ public class Resources extends BasePage {
 
         paginatorForm.add(rowsChooser);
         add(paginatorForm);
+    }
+
+    /**
+     * Check if the delete action is forbidden.
+     *
+     * @param connectorTO object to check
+     * @return true if the action is forbidden, false otherwise
+     */
+    public boolean checkDeleteIsForbidden(ConnInstanceTO connectorTO) {
+        boolean forbidden = false;
+
+        for (ResourceTO resourceTO : resourceRestClient.getAllResources()) {
+            if (resourceTO.getConnectorId().equals(connectorTO.getId())) {
+                forbidden = true;
+            }
+        }
+
+        return forbidden;
     }
 
     class ResourcesProvider extends SortableDataProvider<ResourceTO> {
